pyproject.toml
[project]
name = "ai-customer-support-bot"
version = "0.1.0"
description = "Intelligent AI customer support chatbot with FAQ matching and escalation"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Meet Bhuva"}
]
keywords = ["ai", "chatbot", "customer-support", "fastapi", "llm"]
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi>=0.118.2",
    "google-genai>=1.43.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
[project.scripts]
chat-bot = "main:main"
